+++ linux-syscall/test/testshm1.c
@@ -5,9 +5,8 @@
 #include <stdlib.h>
 #include <string.h>
 #include <time.h>
-#include <sys/types.h>
 #include <sys/shm.h>
-#include <sys/wait.h>
+#include <sys/stat.h>
 #include <unistd.h>
 #include <assert.h>
 #include <stdio.h>
@@ -70,35 +69,19 @@ static void set()
 	T(shmdt(p));
 }
 
-static void get()
-{
-	key_t k;
-	int shmid;
-	void *p;
-
-	T(k = ftok(path, id));
-	T(shmid = shmget(k, 0, 0));
-
-	errno = 0;
-	if ((p=shmat(shmid, 0, SHM_RDONLY)) == 0)
-		printf("shmat failed: %s\n", strerror(errno));
-
-	if (strcmp((char *)p, "test data") != 0)
-		printf("reading shared mem failed: got \"%.100s\" want \"test data\"\n", p);
-
-	/* cleanup */
-	T(shmdt(p));
-	//T(shmctl(shmid, IPC_RMID, 0));
-}
-
 int main(void)
 {
 	int p;
 	int status;
-
-	//set();
-
-	get();
+	set();
+	int pid = vfork();
+	if (pid < 0)
+        printf("error in fork!\n");
+    else if (pid == 0)
+    {
+        execl("/bin/testshm2", "/bin/testshm2", NULL);
+        exit(0);
+    }
 
 	return 0;
 }
